Move tests closer to implementation
